+++ config.js
@@ -1,6 +1,6 @@
 const CONFIG = {
     // App version
-    VERSION: '0.5.3',
+    VERSION: '0.5.4',
     VERSION_NAME: 'Beta (OAuth)',
 
     // Google Sheets API